04_data_transformation/pyproject.toml
[project]
name = "04-data-transformation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-duckdb>=1.10.1",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
]
